feat: 회원 가입 관련 auth 패키지로 이동
