+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -1,0 +1,24 @@
+package woowacourse.auth.ui;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import woowacourse.shoppingcart.application.CustomerService;
+import woowacourse.shoppingcart.dto.SigninRequest;
+import woowacourse.shoppingcart.dto.TokenResponse;
+
+@RestController
+public class AuthController {
+
+    private final CustomerService customerService;
+
+    public AuthController(CustomerService customerService) {
+        this.customerService = customerService;
+    }
+
+    @PostMapping("/signin")
+    public TokenResponse signIn(@RequestBody SigninRequest signinRequest) {
+        return customerService.signIn(signinRequest);
+    }
+
+}
