+++ src/test/java/woowacourse/shoppingcart/acceptance/CustomerAcceptanceTest.java
@@ -304,32 +304,6 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
     }
 
     @Test
-    void 로그인() {
-        String account = "leo8842";
-        String password = "dpepsWkd12!";
-
-        회원이_저장되어_있음(account, password);
-
-        ExtractableResponse<Response> response = 토큰_발급(account, password);
-
-        assertThat(response.statusCode()).isEqualTo(200);
-        assertThat(getValue(response, "accessToken")).isNotNull();
-    }
-
-    @Test
-    void 비밀번호_불일치_로그인_실패() {
-        String account = "leo8842";
-        String password = "dpepsWkd12!";
-
-        회원이_저장되어_있음(account, password);
-
-        ExtractableResponse<Response> response = 토큰_발급(account, "dpepsWkd");
-
-        assertThat(response.statusCode()).isEqualTo(401);
-        assertThat(getValue(response, "message")).contains("로그인이 불가능합니다.");
-    }
-
-    @Test
     void 사용자_정보_조회() {
         String account = "leo8842";
         String password = "dpepsWkd12!";
@@ -584,127 +558,5 @@ public class CustomerAcceptanceTest extends AcceptanceTest {
         // then
         assertThat(response.statusCode()).isEqualTo(400);
         assertThat(getValue(response, "message")).contains("휴대폰 번호양식이 불일치 합니다.");
-    }
-
-
-    private String getValue(ExtractableResponse<Response> response, String value) {
-        return response.body().jsonPath().getString(value);
-    }
-
-    private String getPhoneNumberValue(ExtractableResponse<Response> response, String value) {
-        return String.valueOf(response.body().jsonPath().getMap("phoneNumber").get(value));
-    }
-
-    private Map<String, Object> 회원_정보(String account, String nickname, String password,
-            String address, String start, String middle, String last) {
-        Map<String, Object> request = new HashMap<>();
-        request.put("account", account);
-        request.put("nickname", nickname);
-        request.put("password", password);
-        request.put("address", address);
-        request.put("phoneNumber", 휴대폰_정보(start, middle, last));
-        return request;
-    }
-
-    private Map<String, String> 휴대폰_정보(String start, String middle, String last) {
-        Map<String, String> phoneNumber = new HashMap<>();
-        phoneNumber.put("start", start);
-        phoneNumber.put("middle", middle);
-        phoneNumber.put("last", last);
-        return phoneNumber;
-    }
-
-    private ExtractableResponse<Response> 회원_가입_요청(Map<String, Object> request) {
-        return RestAssured.given()
-                .log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(request)
-                .when()
-                .post("/signup")
-                .then()
-                .log().all()
-                .extract();
-    }
-
-    private ExtractableResponse<Response> 토큰_발급(String account, String password) {
-        Map<String, Object> request = new HashMap<>();
-        request.put("account", account);
-        request.put("password", password);
-
-        return RestAssured.given()
-                .log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(request)
-                .when()
-                .post("/signin")
-                .then()
-                .log().all()
-                .extract();
-    }
-
-    private String 회원_가입_후_토큰_발급(String account, String password) {
-        회원이_저장되어_있음(account, password);
-
-        return getValue(토큰_발급(account, password), "accessToken");
-    }
-
-    private ExtractableResponse<Response> 회원_조회(String accessToken) {
-        ExtractableResponse<Response> response = RestAssured.given()
-                .log().all()
-                .header("Authorization", "Bearer " + accessToken)
-                .when()
-                .get("/customers")
-                .then()
-                .log().all()
-                .extract();
-        return response;
-    }
-
-    private ExtractableResponse<Response> 회원이_저장되어_있음(String account, String password) {
-        return 회원_가입_요청(회원_정보(account,
-                "에덴",
-                password,
-                "에덴 동산",
-                "010",
-                "1234",
-                "5678"));
-    }
-
-    private ExtractableResponse<Response> 회원_탈퇴(String accessToken, String password) {
-        Map<String, Object> request = new HashMap<>();
-        request.put("password", password);
-
-        return RestAssured.given()
-                .log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .header("Authorization", "Bearer" + accessToken)
-                .body(request)
-                .when()
-                .delete("/customers")
-                .then()
-                .log().all()
-                .extract();
-    }
-
-    private Map<String, Object> 회원_수정_정보(String nickname, String address, String start,
-            String middle, String last) {
-        Map<String, Object> request = new HashMap<>();
-        request.put("nickname", nickname);
-        request.put("address", address);
-        request.put("phoneNumber", 휴대폰_정보(start, middle, last));
-        return request;
-    }
-
-    private ExtractableResponse<Response> 회원_수정(String accessToken, Map<String, Object> request) {
-        return RestAssured.given()
-                .log().all()
-                .header("Authorization", "Bearer " + accessToken)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(request)
-                .when()
-                .put("/customers")
-                .then()
-                .log().all()
-                .extract();
     }
 }
